requirements.txt
aiohttp>=3.9.0
py-clob-client>=0.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
web3
py-builder-relayer-client
